Move Events to api, refactor and include shift start task scheduler
