+++ api/src/main/java/org/openmrs/module/ipd/api/events/IPDEventManager.java
@@ -1,9 +1,9 @@
-package org.openmrs.module.ipd.events;
+package org.openmrs.module.ipd.api.events;
 
-import org.openmrs.module.ipd.events.factory.IPDEventFactory;
-import org.openmrs.module.ipd.events.handler.IPDEventHandler;
-import org.openmrs.module.ipd.events.model.IPDEvent;
-import org.openmrs.module.ipd.events.model.IPDEventType;
+import org.openmrs.module.ipd.api.events.model.IPDEvent;
+import org.openmrs.module.ipd.api.events.factory.IPDEventFactory;
+import org.openmrs.module.ipd.api.events.handler.IPDEventHandler;
+import org.openmrs.module.ipd.api.events.model.IPDEventType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
