+++ api/src/main/java/org/openmrs/module/ipd/api/events/factory/IPDEventFactory.java
@@ -1,0 +1,8 @@
+package org.openmrs.module.ipd.api.events.factory;
+
+import org.openmrs.module.ipd.api.events.model.IPDEventType;
+import org.openmrs.module.ipd.api.events.handler.IPDEventHandler;
+
+public interface IPDEventFactory {
+    IPDEventHandler createEventHandler(IPDEventType eventType);
+}
