+++ api/src/main/java/org/openmrs/module/ipd/api/events/handler/IPDEventHandler.java
@@ -1,0 +1,7 @@
+package org.openmrs.module.ipd.api.events.handler;
+
+import org.openmrs.module.ipd.api.events.model.IPDEvent;
+
+public interface IPDEventHandler {
+    void handleEvent(IPDEvent event);
+}
