+++ api/src/main/java/org/openmrs/module/ipd/api/events/handler/impl/PatientAdmitEventHandler.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.ipd.events.handler.impl;
+package org.openmrs.module.ipd.api.events.handler.impl;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -6,12 +6,12 @@ import org.openmrs.module.fhirExtension.model.Task;
 import org.openmrs.module.fhirExtension.service.TaskService;
 import org.openmrs.module.fhirExtension.web.contract.TaskRequest;
 import org.openmrs.module.fhirExtension.web.mapper.TaskMapper;
-import org.openmrs.module.ipd.events.ConfigLoader;
-import org.openmrs.module.ipd.events.IPDEventUtils;
-import org.openmrs.module.ipd.events.handler.IPDEventHandler;
-import org.openmrs.module.ipd.events.model.ConfigDetail;
-import org.openmrs.module.ipd.events.model.IPDEvent;
-import org.openmrs.module.ipd.events.model.TaskDetail;
+import org.openmrs.module.ipd.api.events.ConfigLoader;
+import org.openmrs.module.ipd.api.events.IPDEventUtils;
+import org.openmrs.module.ipd.api.events.model.ConfigDetail;
+import org.openmrs.module.ipd.api.events.model.IPDEvent;
+import org.openmrs.module.ipd.api.events.model.TaskDetail;
+import org.openmrs.module.ipd.api.events.handler.IPDEventHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
