+++ api/src/main/resources/liquibase.xml
@@ -407,7 +407,7 @@
     <changeSet id="shift-start-create-tasks-202404041331" author="Bahmni">
         <preConditions onFail="MARK_RAN">
             <sqlCheck expectedResult="0">
-                SELECT COUNT(*) FROM scheduler_task_config WHERE name = "Mark Slots As Missed";
+                SELECT COUNT(*) FROM scheduler_task_config WHERE name = "Shift Start Tasks";
             </sqlCheck>
         </preConditions>
         <sql>
@@ -415,7 +415,7 @@
             scheduler_task_config(name, description, schedulable_class, start_time, start_time_pattern, repeat_interval,
             start_on_startup, created_by, date_created, uuid)
             VALUES
-            ('Shift Start Tasks', 'Shift Start Tasks', 'org.openmrs.module.ipd.scheduler.ShiftStartTasks',
+            ('Shift Start Tasks', 'Shift Start Tasks', 'org.openmrs.module.ipd.api.scheduler.tasks.ShiftStartTasks',
             TIMESTAMP(CURDATE()), 'MM/dd/yyyy HH:mm:ss', 120, false, 1, NOW(), UUID());
         </sql>
     </changeSet>
