--- omod/src/main/java/org/openmrs/module/ipd/events/factory/IPDEventFactory.java
@@ -1,8 +1,0 @@
-package org.openmrs.module.ipd.events.factory;
-
-import org.openmrs.module.ipd.events.model.IPDEventType;
-import org.openmrs.module.ipd.events.handler.IPDEventHandler;
-
-public interface IPDEventFactory {
-    IPDEventHandler createEventHandler(IPDEventType eventType);
-}
