--- omod/src/main/java/org/openmrs/module/ipd/events/handler/IPDEventHandler.java
@@ -1,7 +1,0 @@
-package org.openmrs.module.ipd.events.handler;
-
-import org.openmrs.module.ipd.events.model.IPDEvent;
-
-public interface IPDEventHandler {
-    void handleEvent(IPDEvent event);
-}
